Major refactoring
Removed a lot of really hard to follow functions and rewrote a bunch of stuff.

There's still at least one bug I'm aware of (it doesn't correctly notice if a comment has been replied to or not), but there are probably more bugs too

Moved a few more things into the config file. Eventually I'd like there to be no references to 'deltas' at all in deltabot.py