Improved force add
Force add now works even if deltabot has already replied. It will not
give a delta if it's already confirmed one, but it will remove any other
comments and re-add them. There is a way to edit instead of delete/add,
but I couldn't get it to work. This should be improved in the future for
aesthetic purposes.

This code assumes that the confirmation message always starts with the
same first 15 characters.