Fixed flair inconsistency
There was a scenario where an account had a delta but it was revoked,
resulting in .get_flair to pass but subsequent functions to fail. I
added a check specifically for flair_text = None (which should only
trigger in these circumstances).

This is most beneficial when testing and constantly resetting an account
to 0 delta points. New accounts should still trigger the "else" portion
of this if statement.