Added history auto-clear
Added "reset_counter" to the go(self): function. This will count to 10.
When it reaches 10, it will clear the comment history and perform a full
scan of the comments. This is a temporary solution to alleviate missed
deltas that we've been seeing lately. A better history feature will
properly solve this issue.